MAE102 Abstracts DeviceOutputBuffer and create DeviceInputBuffer
