+++ PinPadEmulator/Devices/DeviceInputBuffer.cs
@@ -1,0 +1,13 @@
+﻿namespace PinPadEmulator.Devices
+{
+	public class DeviceInputBuffer : DeviceBuffer
+	{
+		public DeviceInputBuffer(IDevice device)
+			: base(device)
+		{
+		}
+
+		public void Write(params byte[] data)
+			=> this.Device.Input(data);
+	}
+}
