+++ PinPadEmulator/Devices/DeviceOutputBuffer.cs
@@ -1,52 +1,20 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Diagnostics;
-using System.Threading;
-
-namespace PinPadEmulator.Devices
+﻿namespace PinPadEmulator.Devices
 {
-	public class DeviceOutputBuffer
+	public class DeviceOutputBuffer : DeviceBuffer
 	{
 		private readonly IDevice device;
 
-		private readonly Queue<byte> buffer = new Queue<byte>();
-		public TimeSpan ReadTimeout = TimeSpan.FromMilliseconds(100);
-
 		public DeviceOutputBuffer(IDevice device)
+			: base(device)
 		{
-			if (device == null) { throw new ArgumentNullException(nameof(device)); }
-
-			this.device = device;
-			this.device.Output += this.OnDeviceOutput;
-		}
-
-		public IEnumerable<byte> Read(int length)
-		{
-			var stopwatch = Stopwatch.StartNew();
-			var counter = 0;
-			while (counter < length && stopwatch.Elapsed < this.ReadTimeout)
-			{
-				lock (this.buffer)
-				{
-					if (this.buffer.Count > 0)
-					{
-						yield return this.buffer.Dequeue();
-						counter++;
-					}
-
-					Thread.Sleep(1);
-				}
-			}
+			this.Device.Output += this.OnDeviceOutput;
 		}
 
 		private void OnDeviceOutput(byte[] output)
 		{
-			lock (this.buffer)
+			lock (this.Buffer)
 			{
-				foreach (var data in output)
-				{
-					this.buffer.Enqueue(data);
-				}
+				foreach (var data in output) { this.Buffer.Enqueue(data); }
 			}
 		}
 	}
